Browser Compatibility: follow standard practices for isomorphic javscript
